Refactor: Reorganize search-related function
